Move functions to a doc_helpers module as the conf.py file was getting out of control
