+++ docs/source/conf.py
@@ -27,21 +27,15 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 
-import glob
 import os
-import shutil
 import subprocess
-import typing
-from urllib.parse import urlparse
-
-from markdown_it import MarkdownIt
-from mdformat.renderer import MDRenderer
-
-if typing.TYPE_CHECKING:
-    from autoapi._objects import PythonObject
-    from markdown_it.token import Token
+import sys
 
 CUR_DIR = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(CUR_DIR)
+
+import doc_helpers  # noqa: E402
+
 DOC_DIR = os.path.dirname(CUR_DIR)
 ROOT_DIR = os.path.dirname(os.path.dirname(CUR_DIR))
 AIQ_DIR = os.path.join(ROOT_DIR, "src", "aiq")
@@ -53,143 +47,15 @@ FILE_URL = f"{PROJECT_URL}/blob/main"
 BUILD_DIR = os.path.join(DOC_DIR, "build")
 API_TREE = os.path.join(BUILD_DIR, "_api_tree")
 
-
-def _create_dir(dest_dir: str):
-    if os.path.exists(dest_dir):
-        shutil.rmtree(dest_dir)
-
-    os.makedirs(dest_dir)
-
-
-def copy_api_tree():
-    # Work-around for https://github.com/readthedocs/sphinx-autoapi/issues/298
-    # AutoAPI support for implicit namespaces is broken, so we need to manually
-    # construct an aiq package with an __init__.py file
-    _create_dir(API_TREE)
-    shutil.copytree(AIQ_DIR, os.path.join(API_TREE, "aiq"))
-    with open(os.path.join(API_TREE, "aiq", "__init__.py"), "w") as f:
-        f.write("")
-
-
 # Copy example Markdown files into the documentation tree
 IGNORE_EXAMPLES = (os.path.join(EXAMPLES_DIR, 'documentation_guides/README.md'), )
-example_readmes = glob.glob(f'{EXAMPLES_DIR}/**/*.md', recursive=True)
-EXAMPLES_INDEX = os.path.join(DOC_EXAMPLES, "index.md")
-DIRECT_LINK_EXTENSIONS = ('.png', '.jpg', '.jpeg', '.gif', '.svg', '.md', '.rst')
 
-
-def url_has_scheme(path: str) -> bool:
-    """Check if the path has a scheme (e.g., http, https)."""
-    parsed_url = urlparse(path)
-    return parsed_url.scheme != ''
-
-
-# re-write links
-def path_updater(doc_path: str, path: str) -> str:
-    # only re-write relative urls without a scheme (https://)
-    if not url_has_scheme(path) and not path.startswith('#'):
-        (_, ext) = os.path.splitext(path)
-        if '/docs/source' in path:
-            # Transform links like `../../docs/source/components/react-agent.md` to `../../components/react-agent.md`
-            path = path.replace('/docs/source', '', 1)
-        elif ext not in DIRECT_LINK_EXTENSIONS:
-            # Re-write links to source code files to point to the GitHub repo. In MD we can link to files in the
-            # source tree directly, but since these are not a part of the documentation, we need to link to the
-            # GitHub repo instead.
-            # First normalize the path
-            if not os.path.isabs(path):
-                dir_name = os.path.dirname(doc_path)
-                norm_path = os.path.normpath(os.path.join(dir_name, path))
-                if os.path.isabs(norm_path):
-                    norm_path = os.path.relpath(norm_path, start=ROOT_DIR)
-
-                path = norm_path
-
-            # replace with github link
-            path = os.path.join(FILE_URL, path)
-
-    return path
-
-
-def token_updater(doc_path: str, token: "Token"):
-    for attr_key in ('href', 'src'):
-        attr = token.attrs.get(attr_key)
-        if attr is not None:
-            # Update the path to remove the '/docs/source' prefix
-            token.attrs[attr_key] = path_updater(doc_path, str(attr))
-
-
-def token_checker(doc_path: str, token: "Token", prefix: str = ''):
-    # The prefix and loc variables are just used for debugging and error reporting
-    loc = f"{prefix}.{token.type}"
-    try:
-        if token.type in ('link_open', 'image'):
-            token_updater(doc_path, token)
-
-        if token.children is not None:
-            for child in token.children:
-                token_checker(doc_path, child, prefix=loc)
-
-    except Exception as e:
-        raise RuntimeError(f"Markdown parsing error at {loc}: {e}") from e
-
-
-def rewrite_markdown(doc_path: str, dest_path: str):
-    """
-    This method serves two purposes:
-        1. It rewrites the markdown file updating links and image paths.
-        2. Performs a copy by writing the updated markdown to the destination path.
-    """
-    md = MarkdownIt()
-
-    with open(doc_path) as f:
-        tokens = md.parse(f.read())
-
-    for token in tokens:
-        token_checker(doc_path, token)
-
-    with open(dest_path, "w") as f:
-        renderer = MDRenderer()
-        f.write(renderer.render(tokens, {}, {}))
-
-
-def copy_examples() -> list[str]:
-    _create_dir(DOC_EXAMPLES)
-
-    destination_docs = []
-    for example_readme in example_readmes:
-        if example_readme in IGNORE_EXAMPLES:
-            continue
-
-        rel_path = os.path.relpath(example_readme, EXAMPLES_DIR)
-        if rel_path == "README.md":
-            # make the top-level README.md file the index
-            dest_path = EXAMPLES_INDEX
-        else:
-            dest_path = os.path.join(DOC_EXAMPLES, rel_path)
-            destination_docs.append(dest_path)
-
-        dest_dir = os.path.dirname(dest_path)
-        os.makedirs(dest_dir, exist_ok=True)
-        rewrite_markdown(example_readme, dest_path)
-
-    return destination_docs
-
-
-def write_examples_index(destination_docs: list[str]):
-    with open(EXAMPLES_INDEX, "a") as f:
-        f.write("\n\n```{toctree}\n:maxdepth: 1\n")
-
-        for doc in destination_docs:
-            relative_path = os.path.relpath(doc, DOC_EXAMPLES)
-            f.write(relative_path + "\n")
-
-        f.write("\n```\n")
-
-
-copy_api_tree()
-destination_docs = copy_examples()
-write_examples_index(destination_docs)
+doc_helpers.copy_api_tree(AIQ_DIR, API_TREE)
+doc_helpers.copy_examples(src_dir=EXAMPLES_DIR,
+                          dest_dir=DOC_EXAMPLES,
+                          root_dir=ROOT_DIR,
+                          github_file_url=FILE_URL,
+                          ignore_files=IGNORE_EXAMPLES)
 
 # -- Project information -----------------------------------------------------
 
@@ -407,24 +273,7 @@ intersphinx_mapping = {
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-default_role
 default_role = "py:obj"
 
-# The defauylt docstring for Pydantic models contains some docstrings that cause parsing warnings for docutils.
-# While this string is tightly tied to a specific version of Pydantic, it is hoped that this will be resolved in future
-# versions of Pydantic.
-PYDANTIC_DEFAULT_DOCSTRING = 'Usage docs: https://docs.pydantic.dev/2.10/concepts/models/\n'
-
-
-def skip_pydantic_special_attrs(app: object, what: str, name: str, obj: "PythonObject", skip: bool,
-                                options: list[str]) -> bool:
-
-    if not skip:
-        bases = getattr(obj, 'bases', [])
-        if (not skip and ('pydantic.BaseModel' in bases or 'EndpointBase' in bases)
-                and obj.docstring.startswith(PYDANTIC_DEFAULT_DOCSTRING)):
-            obj.docstring = ""
-
-    return skip
-
 
 def setup(sphinx):
     # Work-around for for Pydantic docstrings that trigger parsing warnings
-    sphinx.connect("autoapi-skip-member", skip_pydantic_special_attrs)
+    sphinx.connect("autoapi-skip-member", doc_helpers.skip_pydantic_special_attrs)
